File system routes (#96)
This PR fixes #68. It may look huge, but the most part of the PR is the new example directory and tests.

- If there is a directory `pages`, its contents will be used to construct a router. This router will be injected by a plugin into entry object.
- Dynamic routes can be created with ~~`_param.vue` or~~ `[param].vue` (or `:param.vue`, although not on Windows).
- Child routes are created if there are both `some.vue` and `some/` in the same directory. If there's only `some/`, it also works, but adds prefixed routes to common list.
- User-provided entry is no more **required**. Ream will work with nothing but `pages/*.vue` (same as Nuxt). If entry is there, it is used. If there's no pages (and no router that came from entry), an exception will be thrown like it used to be (but better as it is more descriptive).
- It is possible to disable file system routes, change `pages` to something else, or change supported extensions. (I am also thinking about other options, e.g. `trailingSlashes: true`, or a custom filter function.)
- It is possible to disable *entrypoint*. (Not that it's really needed, but it came logically.)

I had to make some small changes in `create-app-template.js` to be able to allow plugins to override the router, and to make entry point optional, but most of the code went into a plugin.

The default naming `pages` (not `routes` and not `views`) was chosen because I believe it reflects better what users are putting there — pages that will show on their site. (Routes are URL matching rules, and views are from MVC world where there are also controllers in front — with `fsRoutes`, users don't create neither of these.)

`write-routes.js` is specifically made to be reusable. In my project, I need to have not one but two file system routers (for project.com and for app.project.com). It is now possible with a small ad-hoc plugin that calls `write-routes.js` twice.

`routes-template.js` is decoupled from Ream API for easier testing.